[refactoring/coupong-19] 4-Layered Architecture 기반 작업
