+++ src/main/java/com/onepage/coupong/business/coupon/dto/CouponEventListDto.java
@@ -1,0 +1,14 @@
+package com.onepage.coupong.business.coupon.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Builder
+@Getter
+public class CouponEventListDto {
+    private String eventName;
+    private String eventCategory;
+    private LocalDateTime startTime;
+}
