--- src/main/java/com/onepage/coupong/coupon/domain/enums/CouponCategory.java
@@ -1,5 +1,0 @@
-package com.onepage.coupong.coupon.domain.enums;
-
-public enum CouponCategory {
-    CHICKEN, PIZZA, COFFEE, HAMBURGER, DEFAULT
-}
