--- src/main/java/com/onepage/coupong/coupon/repository/CouponRepository.java
@@ -1,9 +1,0 @@
-package com.onepage.coupong.coupon.repository;
-
-import com.onepage.coupong.coupon.domain.Coupon;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CouponRepository extends JpaRepository<Coupon, Long> {
-}
