+++ src/main/java/com/onepage/coupong/global/scheduler/CouponEventScheduler.java
@@ -1,9 +1,8 @@
 package com.onepage.coupong.global.scheduler;
 
-import com.onepage.coupong.coupon.domain.CouponEvent;
-import com.onepage.coupong.coupon.domain.enums.CouponCategory;
-import com.onepage.coupong.coupon.service.CouponEventService;
-import lombok.Getter;
+import com.onepage.coupong.jpa.coupon.CouponEvent;
+import com.onepage.coupong.jpa.coupon.enums.CouponCategory;
+import com.onepage.coupong.business.coupon.CouponEventServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
@@ -26,7 +25,7 @@ public class CouponEventScheduler {
     // 각 카테고리별 스케줄 관리
     private Map<CouponCategory, ScheduledFuture<?>> scheduledTasks = new ConcurrentHashMap<>();
 
-    public void scheduleEvent(CouponEvent event, CouponEventService couponEventService) {
+    public void scheduleEvent(CouponEvent event, CouponEventServiceImpl couponEventService) {
         log.info(event.getCategory() + " " + event.getCoupon_publish_nums());
 
         // 이벤트 시작 시간과 종료 시간을 계산
