+++ src/main/java/com/onepage/coupong/jpa/coupon/enums/CouponCategory.java
@@ -1,0 +1,5 @@
+package com.onepage.coupong.jpa.coupon.enums;
+
+public enum CouponCategory {
+    CHICKEN, PIZZA, COFFEE, HAMBURGER, DEFAULT
+}
