+++ src/main/java/com/onepage/coupong/leaderboard/domain/CouponWinningLog.java
@@ -1,6 +1,6 @@
 package com.onepage.coupong.leaderboard.domain;
 
-import com.onepage.coupong.coupon.domain.Coupon;
+import com.onepage.coupong.jpa.coupon.Coupon;
 import com.onepage.coupong.user.domain.User;
 import com.onepage.coupong.leaderboard.domain.enums.WinningCouponState;
 import jakarta.persistence.*;
