+++ src/main/java/com/onepage/coupong/persistence/coupon/CouponEventRepository.java
@@ -1,0 +1,11 @@
+package com.onepage.coupong.persistence.coupon;
+
+import com.onepage.coupong.jpa.coupon.CouponEvent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface CouponEventRepository extends JpaRepository<CouponEvent, Long> {
+    List<CouponEvent> findAllByDateBetween(LocalDateTime startOfDay, LocalDateTime endOfDay);
+}
